refactor(sandbox-first): rename sandbox plugin to sandbox-first
The name "sandbox" was too generic. "sandbox-first" directly conveys
the plugin's philosophy and matches the core skill name.

Renames:
- Plugin directory: plugins/sandbox/ -> plugins/sandbox-first/
- Python package: sandbox_plugin -> sandbox_first
- pyproject project name and script: sandbox-plugin -> sandbox-first
- marketplace.json entry and plugin.json name

Earlier commits on this branch have also been reworded to use the
sandbox-first scope for consistency with marketplace validation.
