+++ plugins/sandbox-first/.claude-plugin/plugin.json
@@ -1,5 +1,5 @@
 {
-  "name": "sandbox",
+  "name": "sandbox-first",
   "description": "Promotes sandbox-first execution by intercepting unnecessary unsandboxed Bash calls",
   "author": {"name": "Josh Nichols", "email": "josh@technicalpickles.com"},
   "repository": "https://github.com/technicalpickles/pickled-claude-plugins",
