+++ plugins/sandbox-first/pyproject.toml
@@ -3,16 +3,16 @@ requires = ["hatchling"]
 build-backend = "hatchling.build"
 
 [project]
-name = "sandbox-plugin"
+name = "sandbox-first"
 version = "0.1.0"
-description = "Claude Code sandbox enforcement plugin"
+description = "Claude Code sandbox-first enforcement plugin"
 requires-python = ">=3.10"
 
 [project.scripts]
-sandbox-plugin = "sandbox_plugin.cli:main"
+sandbox-first = "sandbox_first.cli:main"
 
 [tool.hatch.build.targets.wheel]
-packages = ["src/sandbox_plugin"]
+packages = ["src/sandbox_first"]
 
 [tool.pytest.ini_options]
 testpaths = ["tests"]
