+++ plugins/sandbox-first/skills/sandbox-first/SKILL.md
