+++ plugins/sandbox-first/src/sandbox_first/cli.py
@@ -1,14 +1,14 @@
-"""CLI entry point for sandbox plugin hooks."""
+"""CLI entry point for sandbox-first plugin hooks."""
 
 import json
 import sys
 
-from sandbox_plugin.checker import check_pre_tool_use, check_post_tool_use_failure
+from sandbox_first.checker import check_pre_tool_use, check_post_tool_use_failure
 
 
 def main():
     if len(sys.argv) < 2:
-        print("Usage: sandbox-plugin <pre-tool-use|post-tool-use-failure>", file=sys.stderr)
+        print("Usage: sandbox-first <pre-tool-use|post-tool-use-failure>", file=sys.stderr)
         sys.exit(1)
 
     subcommand = sys.argv[1]
