+++ plugins/sandbox-first/tests/conftest.py
@@ -1,5 +1,5 @@
 import sys
 from pathlib import Path
 
-# Add src to path so tests can import sandbox_plugin
+# Add src to path so tests can import sandbox_first
 sys.path.insert(0, str(Path(__file__).parent.parent / "src"))
